+++ termbox/src/Termbox/Internal/Style.hs
@@ -15,8 +15,8 @@ import Termbox.Bindings.Hs hiding (bg, fg)
 import Termbox.Internal.Color (Color, MaybeColor, justColor, nothingColor, unMaybeColor)
 
 data Style = Style
-  { foreground :: {-# UNPACK #-} !ColorAndAttr,
-    background :: {-# UNPACK #-} !ColorAndAttr
+  { foreground :: {-# UNPACK #-} !ColorAndAttrs,
+    background :: {-# UNPACK #-} !ColorAndAttrs
   }
 
 instance Monoid Style where
@@ -28,27 +28,27 @@ instance Semigroup Style where
   Style a1 b1 <> Style a2 b2 =
     Style (a2 <> a1) (b2 <> b1)
 
--- Render a style as a foreground `tb_color`.
-asForeground :: Style -> Tb_color_and_attrs
+-- Render a style as a foreground `Tb_attrs`.
+asForeground :: Style -> Tb_attrs
 asForeground Style {foreground} =
   renderColorAndAttr foreground
 
--- Render a style as a background `tb_color`.
-asBackground :: Style -> Tb_color_and_attrs
+-- Render a style as a background `Tb_attrs`.
+asBackground :: Style -> Tb_attrs
 asBackground Style {background} =
   renderColorAndAttr background
 
-onlyForeground :: ColorAndAttr -> Style
+onlyForeground :: ColorAndAttrs -> Style
 onlyForeground style =
   mempty {foreground = style}
 
-onlyBackground :: ColorAndAttr -> Style
+onlyBackground :: ColorAndAttrs -> Style
 onlyBackground style =
   mempty {background = style}
 
 maybeFill :: MaybeColor -> Style
 maybeFill color =
-  onlyBackground ColorAndAttr {color, attr = mempty}
+  onlyBackground ColorAndAttrs {color, attrs = mempty}
 
 fg :: Color -> Style
 fg =
@@ -70,30 +70,30 @@ blink :: Style
 blink =
   onlyBackground (onlyAttr _TB_BOLD)
 
-data ColorAndAttr = ColorAndAttr
+data ColorAndAttrs = ColorAndAttrs
   { color :: {-# UNPACK #-} !MaybeColor,
-    attr :: {-# UNPACK #-} !Tb_color_and_attrs
+    attrs :: {-# UNPACK #-} !Tb_attrs
   }
 
-instance Monoid ColorAndAttr where
+instance Monoid ColorAndAttrs where
   mempty =
-    ColorAndAttr nothingColor mempty
+    ColorAndAttrs nothingColor mempty
 
 -- right-biased
-instance Semigroup ColorAndAttr where
-  ColorAndAttr color1 attr1 <> ColorAndAttr color2 attr2 =
-    ColorAndAttr
+instance Semigroup ColorAndAttrs where
+  ColorAndAttrs color1 attr1 <> ColorAndAttrs color2 attr2 =
+    ColorAndAttrs
       (if color2 == nothingColor then color1 else color2)
       (attr1 <> attr2)
 
-renderColorAndAttr :: ColorAndAttr -> Tb_color_and_attrs
-renderColorAndAttr ColorAndAttr {color, attr} =
-  attr <> unMaybeColor color
+renderColorAndAttr :: ColorAndAttrs -> Tb_attrs
+renderColorAndAttr ColorAndAttrs {color, attrs} =
+  attrs <> unMaybeColor color
 
-onlyColor :: Color -> ColorAndAttr
+onlyColor :: Color -> ColorAndAttrs
 onlyColor color =
   mempty {color = justColor color}
 
-onlyAttr :: Tb_color_and_attrs -> ColorAndAttr
-onlyAttr attr =
-  mempty {attr}
+onlyAttr :: Tb_attrs -> ColorAndAttrs
+onlyAttr attrs =
+  mempty {attrs}